+++ docs/changelog.rst
@@ -22,6 +22,7 @@ TODO
 * Compléter si nécessaire les champs ``url``, ``target``, ``picto``, ``groupe`` et ``actif`` dans la table ``synthese.bib_sources``
 * Ajouter le paramètre ``id_source_mortalite = 2;`` au fichier ``web/js/config.js``;
 * Ajouter le paramètre ``public static $id_source_mortalite = 2;`` au fichier ``lib/sfGeonatureConfig.php``;
+* L'altitude est calculée à partir du service "Alticodage" de l'API GeoPortail de l'IGN. Il n'est plus nécessaire de remplir la couche ``layers.l_isolines20``. Cette couche peut toutefois encore être utile si l'utilisateur ne fourni pas l'altitude dans les formulaires de saisie.
 * le loup et le lynx sont retiré de la saisie (saisie recommandée selon protocole du réseau grands prédateurs)
 * le cerf, chamois et le bouquetin doivent être saisie selon 6 critères de sexe et age et non 5 comme les autres taxons. Comportement peut-être changé en modifiant la vue contactfaune.v_nomade_taxons_faune
 * Si vous avez déjà des données de mortalité enregistrées, vous devez adapater la requête SQL ci-dessous avec votre ``id_source`` pour Mortalité et l'exécuter :
@@ -46,6 +47,7 @@ TODO
 * refonte du css du formulaire de login avec bootstrap et une image de fond différente
 * Suppression d'images non utilisées dans le répertoire ``web/images``
 * Mise en cohérence des vues taxonomiques faune #81
+* Calcul de l'altitude à partir du service "Alticodage" de l'API GeoPortail de l'IGN
  
  
 1.4.0 (2015-10-16)